personal-agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "personal-agent"
version = "0.1.0"
description = "A personal AI agent with terminal and web interfaces"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "google-generativeai>=0.3.2",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "rich>=13.7.0",
    "prompt-toolkit>=3.0.41",
    "click>=8.1.7",
    "apscheduler>=3.10.4",
    "redis>=5.0.1",
    "httpx>=0.25.2",
    "aiohttp>=3.9.1",
    "sqlalchemy>=2.0.23",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "ruff>=0.1.6",
]

[project.scripts]
personal-agent = "agent.cli:main"
personal-agent-daemon = "daemon.service:main"

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py310"
